requirements.txt
aiofile==3.9.0
aiopath==0.6.11
aioshutil
alembic
annotated-types
anyio==3.7.1
argon2-cffi
argon2-cffi-bindings
asgiref
bcrypt
black
caio==0.9.17
certifi==2024.12.14
cffi
click
cloudinary==1.42.1
cryptography
Django
dnspython==2.7.0
ecdsa
email_validator==2.2.0
fastapi
fastapi-cache2
fastapi_mail
greenlet
h11
idna
importlib-metadata
Jinja2
Mako
MarkupSafe
mypy-extensions
packaging
passlib==1.7.4
pathspec
pendulum
platformdirs
psycopg2
pyasn1
pycparser
pydantic
pydantic_core
python-dateutil
python-dotenv
python-jose
python-multipart
rsa
six
sniffio
SQLAlchemy
sqlparse
starlette
time-machine
typing_extensions
tzdata
urllib3==2.3.0
uvicorn
zipp
